requirements.txt
openai>=1.40.0
pillow>=10.4.0
python-dotenv>=1.0.1
pydantic>=2.8.2
requests>=2.31.0